rename stepfunctions, optimize imports
