+++ README.md
@@ -34,14 +34,13 @@ from aws_cdk import core
 
 from datajob.datajob_stack import DataJobStack
 from datajob.glue.glue_job import GlueJob
-from datajob.stepfunctions_workflow import StepfunctionsWorkflow
+from datajob.stepfunctions import StepfunctionsWorkflow
 
 app = core.App()
 
 # The datajob_stack is the instance that will result in a cloudformation stack.
 # We inject the datajob_stack object through all the resources that we want to add.
 with DataJobStack(scope=app, id="data-pipeline-simple") as datajob_stack:
-
     # We define 2 glue jobs with the relative path to the source code.
     task1 = GlueJob(
         datajob_stack=datajob_stack, name="task1", job_path="glue_jobs/task.py"
@@ -126,16 +125,15 @@ import pathlib
 from aws_cdk import core
 from datajob.datajob_stack import DataJobStack
 from datajob.glue.glue_job import GlueJob
-from datajob.stepfunctions_workflow import StepfunctionsWorkflow
+from datajob.stepfunctions import StepfunctionsWorkflow
 
 current_dir = str(pathlib.Path(__file__).parent.absolute())
 
 app = core.App()
 
 with DataJobStack(
-    scope=app, id="datajob-python-pyspark", project_root=current_dir
+        scope=app, id="datajob-python-pyspark", project_root=current_dir
 ) as datajob_stack:
-
     pyspark_job = GlueJob(
         datajob_stack=datajob_stack,
         name="pyspark-job",
@@ -228,16 +226,15 @@ import pathlib
 from aws_cdk import core
 from datajob.datajob_stack import DataJobStack
 from datajob.glue.glue_job import GlueJob
-from datajob.stepfunctions_workflow.stepfunctions_workflow import StepfunctionsWorkflow
+from datajob.stepfunctions.stepfunctions_workflow import StepfunctionsWorkflow
 
 current_dir = str(pathlib.Path(__file__).parent.absolute())
 
 app = core.App()
 
 with DataJobStack(
-    scope=app, id="datajob-python-pyspark", project_root=current_dir
+        scope=app, id="datajob-python-pyspark", project_root=current_dir
 ) as datajob_stack:
-
     pyspark_job = GlueJob(
         datajob_stack=datajob_stack,
         name="pyspark-job",
@@ -351,7 +348,7 @@ from aws_cdk import core
 
 from datajob.datajob_stack import DataJobStack
 from datajob.glue.glue_job import GlueJob
-from datajob.stepfunctions_workflow.stepfunctions_workflow import StepfunctionsWorkflow
+from datajob.stepfunctions.stepfunctions_workflow import StepfunctionsWorkflow
 
 current_dir = pathlib.Path(__file__).parent.absolute()
