+++ examples/data_pipeline_simple/datajob_stack_explicit.py
@@ -3,7 +3,7 @@ from aws_cdk import core
 
 from datajob.datajob_stack import DataJobStack
 from datajob.glue.glue_job import GlueJob
-from datajob.stepfunctions_workflow import StepfunctionsWorkflow
+from datajob.stepfunctions import StepfunctionsWorkflow
 
 app = core.App()
