+++ examples/data_pipeline_with_packaged_project/datajob_stack.py
@@ -4,7 +4,7 @@ from aws_cdk import core
 
 from datajob.datajob_stack import DataJobStack
 from datajob.glue.glue_job import GlueJob
-from datajob.stepfunctions_workflow import StepfunctionsWorkflow
+from datajob.stepfunctions import StepfunctionsWorkflow
 
 current_dir = pathlib.Path(__file__).parent.absolute()
