+++ examples/data_pipeline_with_packaged_project/datajob_stack_explicit.py
@@ -5,7 +5,7 @@ from aws_cdk import core
 
 from datajob.datajob_stack import DataJobStack
 from datajob.glue.glue_job import GlueJob
-from datajob.stepfunctions_workflow import StepfunctionsWorkflow
+from datajob.stepfunctions import StepfunctionsWorkflow
 
 app = core.App()
